Move tests from pep526 to type_inference
