+++ tests/run/pep526_variable_annotations.py
@@ -15,7 +15,6 @@ import sys
 
 try:
     import typing
-    from typing import Set as _SET_
     from typing import ClassVar
 except ImportError:
     pass  # this should allow Cython to interpret the directives even when the module doesn't exist
@@ -103,11 +102,6 @@ class BasicStarshipExt(object):
     def __repr__(self):
         return f"BasicStarshipExt(damage={self.damage})"
 
-@cython.cclass
-class AdvancedStarshipExt(BasicStarshipExt):
-    def __repr__(self):
-        return f"AdvancedStarshipExt(damage={self.damage})"
-
 
 T = TypeVar('T')
 
@@ -181,92 +175,6 @@ def literal_list_ptr():
     return a[3]
 
 
-def test_subscripted_types():
-    """
-    >>> test_subscripted_types()
-    dict[int,float] object
-    dict[int,float] object
-    list[int] object
-    list[int] object
-    list object
-    set[Python object] object
-    tuple[int,Python object,float] object
-    tuple[int,Python object,float] object
-    """
-    a1: typing.Dict[cython.int, cython.float] = {}
-    a2: dict[cython.int, cython.float] = {}
-    b1: List[cython.int] = []
-    b2: list[cython.int] = []
-    b3: List = []  # doesn't need to be subscripted
-    c: _SET_[object] = set()
-    t1: Tuple[cython.int, object, cython.float] = (1, 'a', 1.0)
-    t2: tuple[cython.int, object, cython.float] = (1, 'a', 1.0)
-
-
-    print(cython.typeof(a1) + ("[int,float] object" if not cython.compiled else ""))
-    print(cython.typeof(a2) + ("[int,float] object" if not cython.compiled else ""))
-    print(cython.typeof(b1) + ("[int] object" if not cython.compiled else ""))
-    print(cython.typeof(b2) + ("[int] object" if not cython.compiled else ""))
-    print(cython.typeof(b3) + (" object" if not cython.compiled else ""))
-    print(cython.typeof(c) + ("[Python object] object" if not cython.compiled else ""))
-    print(cython.typeof(t1) + ("[int,Python object,float] object" if not cython.compiled else ""))
-    print(cython.typeof(t2) + ("[int,Python object,float] object" if not cython.compiled else ""))
-
-if sys.version_info >= (3, 11):
-    # This part of the test is failing in Python 3.9 and 3.10 with the following exception in Shadow.py:
-    # isinstance() argument 2 cannot be a parameterized generic
-
-    def test_casting_subscripted_types():
-        """
-        >>> test_casting_subscripted_types()
-        tuple:
-        tuple[list object,int,int] object 1 2
-        tuple object 1.0 2.0
-        list:
-        list[int] object 1
-        list object 1.0
-        dict:
-        dict[str object,float] object 2.0
-        dict object 2
-        float 3.0
-        Python object 3
-        """
-        print('tuple:')
-        t: tuple[list, cython.float, cython.float] = ([], 1.0, 2.0)
-        x0 = cython.cast(tuple[list, cython.int, cython.int], t)
-        y0 = cython.cast(tuple, t)
-        print(
-            cython.typeof(x0) + ('[list object,int,int] object' if not cython.compiled else ''),
-            x0[1] if cython.compiled else int(x0[1]),
-            x0[2] if cython.compiled else int(x0[2]),
-        )
-        print(cython.typeof(y0) + (' object' if not cython.compiled else ''), y0[1], y0[2])
-        print('list:')
-        l: list[cython.float] = [1.0, 2.0]
-        x1 = cython.cast(list[cython.int], l)
-        y1 = cython.cast(list, l)
-        print(cython.typeof(x1) + ('[int] object' if not cython.compiled else ""), int(x1[0]) if not cython.compiled else x1[0])
-        print(cython.typeof(y1) + (' object' if not cython.compiled else ""), y1[0])
-        print('dict:')
-        d: dict[str, cython.int] = {'a': 1, 'b': 2}
-        x2 = cython.cast(dict[str, cython.float], d)
-        y2 = cython.cast(dict, d)
-        print(
-            cython.typeof(x2) + ('[str object,float] object' if not cython.compiled else ""),
-            x2['b'] if cython.compiled else float(x2['b'])
-        )
-        print(cython.typeof(y2) + (' object' if not cython.compiled else ""), y2['b'])
-
-        d2: dict[cython.int, str] = {3: '3'}
-        for k1 in cython.cast(dict[cython.float, str], d2):
-            print(
-                cython.typeof(k1) if cython.compiled else 'float',
-                k1 if cython.compiled else float(k1)
-            )
-
-        for k2 in cython.cast(dict, d2):
-            print(cython.typeof(k2) if cython.compiled else 'Python object', k2)
-
 def test_use_typing_attributes_as_non_annotations():
     """
     >>> test_use_typing_attributes_as_non_annotations()
@@ -315,293 +223,6 @@ def test_use_typing_attributes_as_non_annotations():
     print(name_of(q1), str(q2) in ["typing.Union[typing.FrozenSet, NoneType]", "typing.FrozenSet | None"] or str(q2))
     print(name_of(w1), str(w2) in ["typing.Union[typing.Dict, NoneType]", "typing.Dict | None"] or str(w2))
 
-def test_tuple_with_int_str_subscript():
-    """
-    >>> test_tuple_with_int_str_subscript()
-    str object
-    int object
-    str object
-    int object
-    str object
-    int object
-    FooBar
-    3
-    """
-    a: tuple[str, int] = ("Foo", 1)
-    b: Tuple[str, int] = ("Bar", 2)
-    print(cython.typeof(a[0]) + (" object" if not cython.compiled else ""))
-    print(cython.typeof(a[1]) + (" object" if not cython.compiled else ""))
-    print(cython.typeof(b[0]) + (" object" if not cython.compiled else ""))
-    print(cython.typeof(b[1]) + (" object" if not cython.compiled else ""))
-    print(cython.typeof(a[0] + b[0]) + (" object" if not cython.compiled else ""))
-    print(cython.typeof(a[1] + b[1]) + (" object" if not cython.compiled else ""))
-    print(a[0] + b[0])
-    print(a[1] + b[1])
-
-def test_list_with_str_subscript():
-    """
-    >>> test_list_with_str_subscript()
-    str object
-    str object
-    str object
-    FooBar
-    """
-    a: list[str] = ["Foo"]
-    b: List[str] = ["Bar"]
-    print(cython.typeof(a[0]) + (" object" if not cython.compiled else ""))
-    print(cython.typeof(b[0]) + (" object" if not cython.compiled else ""))
-    print(cython.typeof(a[0] + b[0]) + (" object" if not cython.compiled else ""))
-    print(a[0] + b[0])
-
-def test_list_with_int_subscript():
-    """
-    >>> test_list_with_int_subscript()
-    int
-    int
-    int
-    int
-    3
-    """
-    a: List[cython.int] = [1]
-    b: list[cython.int] = [2]
-    c = a[0] + b[0]
-    print(cython.typeof(a[0]))
-    print(cython.typeof(b[0]))
-    print(cython.typeof(a[0] + b[0]))
-    print(cython.typeof(c))
-    print(c)
-
-def test_dict_with_subscript():
-    """
-    >>> test_dict_with_subscript()
-    int
-    int
-    int
-    int
-    3
-    """
-    a: Dict[str, cython.int] = {"a": 1}
-    b: Dict[cython.int, cython.int] = {1: 2}
-    c = a["a"] + b[1]
-    print(cython.typeof(a["a"]))
-    print(cython.typeof(b[1]))
-    print(cython.typeof(a["a"] + b[1]))
-    print(cython.typeof(c))
-    print(c)
-
-def test_assignment_tuple_with_subscript():
-    """
-    >>> test_assignment_tuple_with_subscript()
-    int int
-    Python object Python object
-    float float
-    5 5 5.0
-    6 6 6.0
-    """
-    a: tuple[cython.int, cython.int] = (5, 6)
-    b: tuple = a
-    c: tuple[cython.float, cython.float] = b
-    print(cython.typeof(a[0]), cython.typeof(a[1])
-    )
-    print(
-        cython.typeof(b[0]) if cython.compiled else "Python object",
-        cython.typeof(b[1]) if cython.compiled else "Python object"
-    )
-    print(
-        cython.typeof(c[0]) if cython.compiled else "float",
-        cython.typeof(c[1]) if cython.compiled else "float"
-    )
-    print(a[0], b[0], c[0] if cython.compiled else float(c[0]))
-    print(a[1], b[1], c[1] if cython.compiled else float(c[1]))
-
-def test_assignment_list_with_subscript():
-    """
-    >>> test_assignment_list_with_subscript()
-    int
-    Python object
-    float
-    5 5 5.0
-    """
-    a: list[cython.int] = [5]
-    b: list = a
-    c: list[cython.float] = b
-    print(cython.typeof(a[0]))
-    print(cython.typeof(b[0]) if cython.compiled else 'Python object')
-    print(cython.typeof(c[0]) if cython.compiled else 'float')
-    print(a[0], b[0], c[0] if cython.compiled else float(c[0]))
-
-def test_assignment_dict_with_subscript():
-    """
-    >>> test_assignment_dict_with_subscript()
-    int
-    Python object
-    float
-    5 5 5.0
-    """
-    a: dict[str, cython.int] = {'a': 5}
-    b: dict = a
-    c: dict[str, cython.float] = b
-    print(cython.typeof(a['a']) if cython.compiled else 'int')
-    print(cython.typeof(b['a']) if cython.compiled else 'Python object')
-    print(cython.typeof(c['a']) if cython.compiled else 'float')
-    print(a['a'], b['a'], c['a'] if cython.compiled else float(c['a']))
-
-if sys.version_info >= (3, 15) or cython.compiled:
-
-    def test_assignment_frozendict_with_subscript():
-        """
-        >>> test_assignment_frozendict_with_subscript()
-        int
-        Python object
-        float
-        5 5 5.0
-        """
-        a: frozendict[str, cython.int] = frozendict({'a': 5})
-        b: frozendict = a
-        c: frozendict[str, cython.float] = b
-        print(cython.typeof(a['a']))
-        print(cython.typeof(b['a']) if cython.compiled else 'Python object')
-        print(cython.typeof(c['a']) if cython.compiled else 'float')
-        print(a['a'], b['a'], c['a'] if cython.compiled else float(c['a']))
-
-def test_failed_assignment_tuple_with_subscript():
-    """
-    >>> test_failed_assignment_tuple_with_subscript()
-    5 5 5
-    """
-    a: tuple[cython.int, cython.int] = (5, 5)
-    b: tuple = a
-    c: tuple[str, str] = b
-    print(a[0], b[0], c[0])
-
-if cython.compiled:
-    test_failed_assignment_tuple_with_subscript.__doc__ = """
-    >>> test_failed_assignment_tuple_with_subscript()  #doctest: +ELLIPSIS
-    Traceback (most recent call last):
-       ...
-    TypeError: Expected str, got int
-    """
-
-def test_failed_assignment_list_with_subscript():
-    """
-    >>> test_failed_assignment_list_with_subscript()
-    5 5 5
-    """
-    a: list[cython.int] = [5]
-    b: list = a
-    c: list[str] = b
-    print(a[0], b[0], c[0])
-
-if cython.compiled:
-    test_failed_assignment_list_with_subscript.__doc__ = """
-    >>> test_failed_assignment_list_with_subscript()  #doctest: +ELLIPSIS
-    Traceback (most recent call last):
-       ...
-    TypeError: Expected str, got int
-    """
-
-@cython.infer_types(True)
-def test_iteration_over_tuple_with_subscript():
-    """
-    >>> test_iteration_over_tuple_with_subscript()
-    2 int
-    bar str
-    BasicStarshipExt(damage=5) BasicStarshipExt
-    AdvancedStarshipExt(damage=6) AdvancedStarshipExt
-    """
-    a: tuple[cython.int, str] = (2, 'bar')
-    for c in a:
-        print(c, cython.typeof(c))
-    # Iterating must use spanning type, not subscripted type, so that it works even when the tuple is heterogeneous.
-    b: tuple[BasicStarshipExt, AdvancedStarshipExt] = (BasicStarshipExt(5), AdvancedStarshipExt(6))
-    for s in b:
-        print(s, cython.typeof(s))
-
-if cython.compiled:
-    test_iteration_over_tuple_with_subscript.__doc__ = """
-    >>> test_iteration_over_tuple_with_subscript()
-    2 Python object
-    bar Python object
-    BasicStarshipExt(damage=5) BasicStarshipExt
-    AdvancedStarshipExt(damage=6) BasicStarshipExt
-    """
-
-@cython.infer_types(True)
-def test_iteration_over_list_with_subscript():
-    """
-    >>> test_iteration_over_list_with_subscript()
-    int
-    int
-    3
-    """
-    b: cython.int = 1
-    a: list[cython.int] = [2]
-    for c in a:
-        print(cython.typeof(c))
-        print(cython.typeof(b + c))
-        print(b + c)
-
-@cython.infer_types(True)
-def test_iteration_over_set_with_subscript():
-    """
-    >>> test_iteration_over_set_with_subscript()
-    int
-    int
-    3
-    """
-    b: cython.int = 1
-    a: set[cython.int] = {2}
-    for c in a:
-        print(cython.typeof(c))
-        print(cython.typeof(b + c))
-        print(b + c)
-
-@cython.infer_types(True)
-def test_iteration_over_frozenset_with_subscript():
-    """
-    >>> test_iteration_over_frozenset_with_subscript()
-    int
-    int
-    3
-    """
-    b: cython.int = 1
-    a: frozenset[cython.int] = frozenset({2})
-    for c in a:
-        print(cython.typeof(c))
-        print(cython.typeof(b + c))
-        print(b + c)
-
-@cython.infer_types(True)
-def test_iteration_over_dict_with_subscript():
-    """
-    >>> test_iteration_over_dict_with_subscript()
-    int
-    int
-    3
-    """
-    b: cython.int = 1
-    a: dict[cython.int, cython.str] = {2: 'a'}
-    for c in a:
-        print(cython.typeof(c))
-        print(cython.typeof(b + c))
-        print(b + c)
-
-if sys.version_info >= (3, 15) or cython.compiled:
-
-    @cython.infer_types(True)
-    def test_iteration_over_frozendict_with_subscript():
-        """
-        >>> test_iteration_over_frozendict_with_subscript()
-        int
-        int
-        3
-        """
-        b: cython.int = 1
-        a: frozendict[cython.int, cython.str] = frozendict({2: 'a'})
-        for c in a:
-            print(cython.typeof(c))
-            print(cython.typeof(b + c))
-            print(b + c)
 
 try:
     import numpy.typing as npt
