+++ tests/run/type_inference_generics.py
@@ -4,12 +4,24 @@
 import sys
 import cython
 
-from typing import Dict, List
+from typing import Dict, List, Tuple
 try:
     import typing
     from typing import Set as _SET_
 except ImportError:
     pass  # this should allow Cython to interpret the directives even when the module doesn't exist
+
+
+@cython.cclass
+class A:
+    def __str__(self):
+        return "A()"
+
+
+@cython.cclass
+class B(A):
+    def __str__(self):
+        return "B()"
 
 
 def min_plain_pyversion(major, minor):
@@ -28,6 +40,8 @@ def test_subscripted_types():
     list[int] object
     list object
     set[Python object] object
+    tuple[int,Python object,float] object
+    tuple[int,Python object,float] object
     """
     a1: typing.Dict[cython.int, cython.float] = {}
     a2: dict[cython.int, cython.float] = {}
@@ -35,6 +49,9 @@ def test_subscripted_types():
     b2: list[cython.int] = []
     b3: List = []  # doesn't need to be subscripted
     c: _SET_[object] = set()
+    t1: Tuple[cython.int, object, cython.float] = (1, 'a', 1.0)
+    t2: tuple[cython.int, object, cython.float] = (1, 'a', 1.0)
+
 
     print(cython.typeof(a1) + ("[int,float] object" if not cython.compiled else ""))
     print(cython.typeof(a2) + ("[int,float] object" if not cython.compiled else ""))
@@ -42,6 +59,8 @@ def test_subscripted_types():
     print(cython.typeof(b2) + ("[int] object" if not cython.compiled else ""))
     print(cython.typeof(b3) + (" object" if not cython.compiled else ""))
     print(cython.typeof(c) + ("[Python object] object" if not cython.compiled else ""))
+    print(cython.typeof(t1) + ("[int,Python object,float] object" if not cython.compiled else ""))
+    print(cython.typeof(t2) + ("[int,Python object,float] object" if not cython.compiled else ""))
 
 
 @min_plain_pyversion(3, 11)
@@ -50,20 +69,35 @@ def test_subscripted_types():
 def test_casting_subscripted_types():
     """
     >>> test_casting_subscripted_types()
+    tuple:
+    tuple[list object,int,int] object 1 2
+    tuple object 1.0 2.0
+    list:
     list[int] object 1
     list object 1.0
+    dict:
     dict[str object,float] object 2.0
     dict object 2
     float 3.0
     Python object 3
     """
-    # list
+    print('tuple:')
+    t: tuple[list, cython.float, cython.float] = ([], 1.0, 2.0)
+    x0 = cython.cast(tuple[list, cython.int, cython.int], t)
+    y0 = cython.cast(tuple, t)
+    print(
+        cython.typeof(x0) + ('[list object,int,int] object' if not cython.compiled else ''),
+        x0[1] if cython.compiled else int(x0[1]),
+        x0[2] if cython.compiled else int(x0[2]),
+    )
+    print(cython.typeof(y0) + (' object' if not cython.compiled else ''), y0[1], y0[2])
+    print('list:')
     l: list[cython.float] = [1.0, 2.0]
     x1 = cython.cast(list[cython.int], l)
     y1 = cython.cast(list, l)
     print(cython.typeof(x1) + ('[int] object' if not cython.compiled else ""), int(x1[0]) if not cython.compiled else x1[0])
     print(cython.typeof(y1) + (' object' if not cython.compiled else ""), y1[0])
-    # dict
+    print('dict:')
     d: dict[str, cython.int] = {'a': 1, 'b': 2}
     x2 = cython.cast(dict[str, cython.float], d)
     y2 = cython.cast(dict, d)
@@ -82,6 +116,55 @@ def test_casting_subscripted_types():
 
     for k2 in cython.cast(dict, d2):
         print(cython.typeof(k2) if cython.compiled else 'Python object', k2)
+
+
+def test_tuple_with_int_str_subscript():
+    """
+    >>> test_tuple_with_int_str_subscript()
+    str object
+    int object
+    str object
+    int object
+    str object
+    int object
+    FooBar
+    3
+    """
+    a: tuple[str, int] = ("Foo", 1)
+    b: Tuple[str, int] = ("Bar", 2)
+    print(cython.typeof(a[0]) + (" object" if not cython.compiled else ""))
+    print(cython.typeof(a[1]) + (" object" if not cython.compiled else ""))
+    print(cython.typeof(b[0]) + (" object" if not cython.compiled else ""))
+    print(cython.typeof(b[1]) + (" object" if not cython.compiled else ""))
+    print(cython.typeof(a[0] + b[0]) + (" object" if not cython.compiled else ""))
+    print(cython.typeof(a[1] + b[1]) + (" object" if not cython.compiled else ""))
+    print(a[0] + b[0])
+    print(a[1] + b[1])
+
+def test_assignment_tuple_with_subscript():
+    """
+    >>> test_assignment_tuple_with_subscript()
+    int int
+    Python object Python object
+    float float
+    5 5 5.0
+    6 6 6.0
+    """
+    a: tuple[cython.int, cython.int] = (5, 6)
+    b: tuple = a
+    c: tuple[cython.float, cython.float] = b
+    print(cython.typeof(a[0]), cython.typeof(a[1])
+    )
+    print(
+        cython.typeof(b[0]) if cython.compiled else "Python object",
+        cython.typeof(b[1]) if cython.compiled else "Python object"
+    )
+    print(
+        cython.typeof(c[0]) if cython.compiled else "float",
+        cython.typeof(c[1]) if cython.compiled else "float"
+    )
+    print(a[0], b[0], c[0] if cython.compiled else float(c[0]))
+    print(a[1], b[1], c[1] if cython.compiled else float(c[1]))
 
 
 def test_list_with_str_subscript():
@@ -190,6 +273,70 @@ def test_assignment_frozendict_with_subscript():
     print(a['a'], b['a'], c['a'] if cython.compiled else float(c['a']))
 
 
+def test_failed_assignment_tuple_with_subscript():
+    """
+    >>> test_failed_assignment_tuple_with_subscript()
+    5 5 5
+    """
+    a: tuple[cython.int, cython.int] = (5, 5)
+    b: tuple = a
+    c: tuple[str, str] = b
+    print(a[0], b[0], c[0])
+
+if cython.compiled:
+    test_failed_assignment_tuple_with_subscript.__doc__ = """
+    >>> test_failed_assignment_tuple_with_subscript()  #doctest: +ELLIPSIS
+    Traceback (most recent call last):
+       ...
+    TypeError: Expected str, got int
+    """
+
+
+def test_failed_assignment_list_with_subscript():
+    """
+    >>> test_failed_assignment_list_with_subscript()
+    5 5 5
+    """
+    a: list[cython.int] = [5]
+    b: list = a
+    c: list[str] = b
+    print(a[0], b[0], c[0])
+
+if cython.compiled:
+    test_failed_assignment_list_with_subscript.__doc__ = """
+    >>> test_failed_assignment_list_with_subscript()  #doctest: +ELLIPSIS
+    Traceback (most recent call last):
+       ...
+    TypeError: Expected str, got int
+    """
+
+@cython.infer_types(True)
+def test_iteration_over_tuple_with_subscript():
+    """
+    >>> test_iteration_over_tuple_with_subscript()
+    2 int
+    bar str
+    A() A
+    B() B
+    """
+    a: tuple[cython.int, str] = (2, 'bar')
+    for c in a:
+        print(c, cython.typeof(c))
+    # Iterating must use spanning type, not subscripted type, so that it works even when the tuple is heterogeneous.
+    b: tuple[A, B] = (A(), B())
+    for s in b:
+        print(s, cython.typeof(s))
+
+if cython.compiled:
+    test_iteration_over_tuple_with_subscript.__doc__ = """
+    >>> test_iteration_over_tuple_with_subscript()
+    2 Python object
+    bar Python object
+    A() A
+    B() A
+    """
+
+
 @cython.infer_types(True)
 def test_iteration_over_list_with_subscript():
     """
@@ -220,6 +367,7 @@ def test_iteration_over_set_with_subscript():
         print(cython.typeof(c))
         print(cython.typeof(b + c))
         print(b + c)
+
 
 @cython.infer_types(True)
 def test_iteration_over_frozenset_with_subscript():
@@ -414,16 +562,6 @@ def test_inference_of_frozendict_constructor():
     print(cython.typeof(fdlls) + ("[str object,str object] object" if not cython.compiled else ""))
     print(cython.typeof(fdll) + (" object" if not cython.compiled else ""))
     print(cython.typeof(fdl) + (" object" if not cython.compiled else ""))
-
-
-@cython.cclass
-class A:
-    pass
-
-
-@cython.cclass
-class B(A):
-    pass
 
 
 def test_inference_of_tuple_common_subscripted_type():
